First part of a package-level add on to the kernel additions
Here, just the existing code wirten by @davidcarlisle for recovering
register numbers is added to LaTeX. This is reordered somewhat from
his original.

The package name is currently the rather wordy "lualatexsupport" mainly
as this is likely to avoid any clashes.

How to handle interaction with "luatex"/"luatexbase", including any
fall-back when used with older kernels, is not yet covered at all.