Add module management functions
These are more-or-less copies from luatexbase. However, it seems
most sensible to stick to the syntax used by LaTeX in
\ProvidesPackage for latex.provides_module. That means we have
two arguments, one of which is just the name, rather than a table
where one entry is required.

To follow is a function to recover module info: seems sensible here.